chore(configs): move defaultConfigs to configs
