+++ packages/configs/src/configs/__tests__/defaultConfigs.js
